update install script

+++ scripts/update.sh
@@ -19,10 +19,12 @@ cp -r bedrock_server bedrock_server_symbols.debug behavior_packs/ definitions/ r
 rm -r /tmp/bedrock_server /tmp/bedrock_server-latest.zip
 
 # download latest scripts into tmp directory, unzip and copy into directory
-echo "Updating BDS-Utils Scripts..."
-wget -O /tmp/bds-utils_scripts-only.zip https://github.com/bmar116/bds-utils/releases/latest/download/bds-utils_scripts-only.zip && unzip -d /tmp/bds-utils_scripts-only/ /tmp/bds-utils_scripts-only.zip
-cp -r /tmp/bds-utils_scripts-only/* ${SERVER_ROOT}/${SCRIPT_DIR}/
-rm -r /tmp/bds-utils_scripts-only/ /tmp/bds-utils_scripts-only.zip
+if [ ! $1 ]; then
+	echo "Updating BDS-Utils Scripts..."
+	wget -O /tmp/bds-utils_scripts-only.zip https://github.com/bmar116/bds-utils/releases/latest/download/bds-utils_scripts-only.zip && unzip -d /tmp/bds-utils_scripts-only/ /tmp/bds-utils_scripts-only.zip
+	cp -r /tmp/bds-utils_scripts-only/* ${SERVER_ROOT}/${SCRIPT_DIR}/
+	rm -r /tmp/bds-utils_scripts-only/ /tmp/bds-utils_scripts-only.zip
+fi
 
 # restart server software
 screen -L -Logfile ${SERVER_ROOT}/${LOG_DIR}/latest.log -dmS ${SCREEN_NAME} bash ${SERVER_ROOT}/${SERVER_DIR}/start.sh ${SERVER_ROOT}/${SCRIPT_DIR}
